feat: add local tracking for scope

+++ engine/src/vm/call_frame.h
@@ -9,7 +9,7 @@ namespace yatsi {
 struct CallFrame {
   BytecodeFunction* function;
   uint32_t ip = 0;        // Instruction pointer (index into function->code)
-  uint8_t base_register;  // Offset into the VM's flat register file
+  uint16_t base_register;  // Offset into the VM's flat register file
 };
 
 } // namespace yatsi
